--- core/navigation/src/main/java/com/yapp/breake/core/navigation/Route.kt
@@ -1,3 +1,0 @@
-package com.yapp.breake.core.navigation
-
-sealed interface Route
