+++ core/navigation/src/main/java/com/yapp/breake/core/navigation/route/Route.kt
@@ -1,0 +1,3 @@
+package com.yapp.breake.core.navigation.route
+
+sealed interface Route
